+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version [6.1.0](https://github.com/cedx/coveralls.js/compare/v6.0.0...v6.1.0)
 - Ported the source code to [TypeScript](https://www.typescriptlang.org).
+- Added the `eventRequest` and `eventResponse` static properties to the `Client` class.
 - Updated the package dependencies.
 
 ## Version [6.0.0](https://github.com/cedx/coveralls.js/compare/v5.1.0...v6.0.0)